Add files via upload
The first project is an age-based quiz game designed to cater to users of different age groups by customizing the questions accordingly. Upon launching the game, the user is prompted to enter their age, which dynamically adjusts the set of questions to be age-appropriate and relevant. This ensures a tailored and educational experience for each participant. The game tests the user's knowledge across various topics, and their performance is tracked throughout the quiz. At the conclusion of the quiz, a total score is calculated, and participants are rewarded with a digital certificate of participation, which includes their score. This feature not only provides a sense of achievement but also encourages engagement and learning. This project demonstrates my ability to implement conditional logic, manage user input, and generate output in a user-friendly manner.

The second project, "Guess the Fortune," is designed to add an element of fun to the workplace. This game invites employees to enter a number, which serves as a key to unlock their unique fortune for the day. The application uses this input to generate a randomized, motivational schedule, predicting various aspects of the employee's day at work. By incorporating randomization and user input handling, this project brings a lighthearted and entertaining twist to daily routines