Add files via upload
Guess The fortune : COMPLETE FILE


"Guess the Fortune," is designed to add an element of fun to the workplace. This game invites employees to enter a number, which serves as a key to unlock their unique fortune for the day. The application uses this input to generate a randomized, motivational schedule, predicting various aspects of the employee's day at work. By incorporating randomization and user input handling, this project brings a lighthearted and entertaining twist to daily routines